refactor: unify rate limiter stores behind a generic seam
Replace two structurally identical token-bucket store implementations
(rateLimiterStore keyed by uuid.UUID and tokenRateLimiterStore keyed by
string) with a single generic rateLimiterStore[K comparable]. The shared
getLimiter and cleanupStale logic now lives once; the two public middleware
functions become thin adapters that supply a key extractor, a log
attribute builder, and their message strings to a shared
newRateLimitMiddleware factory.
